// src/api/api.ts
const BASE_PATH = "/".replace(/\/+$/, "");

/**
 *
 * @export
 */
export const COLLECTION_FORMATS = {
    csv: ",",
    ssv: " ",
    tsv: "\t",
    pipes: "|",
};

/**
 *
 * @export
 * @interface FetchAPI
 */
export interface FetchAPI {
    (url: string, init?: any): Promise<Response>;
}

/**
 *
 * @export
 * @interface FetchArgs
 */
export interface FetchArgs {
    url: string;
    options: any;
}

/**
 *
 * @export
 * @class BaseAPI
 */
export class BaseAPI {
    protected configuration: Configuration;

    constructor(configuration?: Configuration, protected basePath: string = BASE_PATH, protected fetch: FetchAPI = isomorphicFetch) {
        if (configuration) {
            this.configuration = configuration;
            this.basePath = configuration.basePath || this.basePath;
        }
    }
};

/**
 *
 * @export
 * @class RequiredError
 * @extends {Error}
 */
export class RequiredError extends Error {
    name: "RequiredError"
    constructor(public field: string, msg?: string) {
        super(msg);
    }
}

/**
 * 
 * @export
 * @interface Attendance
 */
export interface Attendance {
    /**
     * 
     * @type {number}
     * @memberof Attendance
     */
    attendanceId: number;
    /**
     * 
     * @type {number}
     * @memberof Attendance
     */
    parentAttendanceId?: number;
    /**
     * 
     * @type {number}
     * @memberof Attendance
     */
    staffId: number;
    /**
     * 
     * @type {string}
     * @memberof Attendance
     */
    workDate: string;
    /**
     * 
     * @type {string}
     * @memberof Attendance
     */
    startTime: string;
    /**
     * 
     * @type {string}
     * @memberof Attendance
     */
    endTime?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Attendance
     */
    goDirectlyFlag?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Attendance
     */
    returnDirectlyFlag?: boolean;
    /**
     * 
     * @type {string}
     * @memberof Attendance
     */
    remarks?: string;
}
/**
 * 
 * @export
 * @interface AttendanceClockIn
 */
export interface AttendanceClockIn {
    /**
     * 
     * @type {string}
     * @memberof AttendanceClockIn
     */
    startTime: string;
    /**
     * 
     * @type {boolean}
     * @memberof AttendanceClockIn
     */
    goDirectlyFlag: boolean;
}
/**
 * 
 * @export
 * @interface AttendanceClockOut
 */
export interface AttendanceClockOut {
    /**
     * 
     * @type {string}
     * @memberof AttendanceClockOut
     */
    endTime: string;
    /**
     * 
     * @type {boolean}
     * @memberof AttendanceClockOut
     */
    returnDirectlyFlag: boolean;
}
/**
 * 
 * @export
 * @interface AttendanceRemarks
 */
export interface AttendanceRemarks {
    /**
     * 
     * @type {string}
     * @memberof AttendanceRemarks
     */
    remarks: string;
}
// ...
export class AttendanceApi extends BaseAPI {
    /**
     * 指定したIDの勤怠情報を取得します。
     * @summary スタッフIDで勤怠情報を取得
     * @param {number} staffId スタッフID
     * @param {number} workDate 勤務日
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AttendanceApi
     */
    public readAttendancesAttendancesStaffIdWorkDateGet(staffId: number, workDate: number, options?: any) {
        return AttendanceApiFp(this.configuration).readAttendancesAttendancesStaffIdWorkDateGet(staffId, workDate, options)(this.fetch, this.basePath);
    }

    /**
     * 出勤方法は、`通常`と`直行`の2つ方法で登録することができます。<br>         <br>         **通常**<br>         通常でリクエストする場合は、`start_time`を指定してください。<br>         `go_directly_flag`は自動で`false`として処理されます。<br>         <br>         **直行**<br>         直行でリクエストする場合は、`go_directly_flag`を`true`にしてください。<br>         `start_time`は自動で`09:00:00`として処理されます。<br>
     * @summary 出勤
     * @param {AttendanceClockIn} body 
     * @param {number} staffId スタッフID
     * @param {number} workDate 勤務日
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AttendanceApi
     */
    public registerClockInAttendancesStaffIdWorkDateClockInPost(body: AttendanceClockIn, staffId: number, workDate: number, options?: any) {
        return AttendanceApiFp(this.configuration).registerClockInAttendancesStaffIdWorkDateClockInPost(body, staffId, workDate, options)(this.fetch, this.basePath);
    }

    /**
     * 退勤方法は、`通常`と`直帰`の2つの方法で登録することができます。<br>         <br>         **通常**<br>         通常でリクエストする場合は、`end_time`を指定してください。<br>         `return_directly_flag`は自動で`false`として処理されます。<br>         <br>         **直帰**<br>         直帰でリクエストする場合は、`return_direct_flag`を`true`にしてください。<br>         `start_time`は自動で`09:00:00`として処理されます。<br>
     * @summary 退勤時刻を登録
     * @param {AttendanceClockOut} body 
     * @param {number} staffId スタッフID
     * @param {number} workDate 勤務日
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AttendanceApi
     */
    public registerClockOutAttendancesStaffIdWorkDateClockOutPatch(body: AttendanceClockOut, staffId: number, workDate: number, options?: any) {
        return AttendanceApiFp(this.configuration).registerClockOutAttendancesStaffIdWorkDateClockOutPatch(body, staffId, workDate, options)(this.fetch, this.basePath);
    }

    /**
     * 備考を登録します。<br>
     * @summary 備考を登録
     * @param {AttendanceRemarks} body 
     * @param {number} staffId スタッフID
     * @param {number} workDate 勤務日
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AttendanceApi
     */
    public registerRemarksAttendancesStaffIdWorkDateRemarksPatch(body: AttendanceRemarks, staffId: number, workDate: number, options?: any) {
        return AttendanceApiFp(this.configuration).registerRemarksAttendancesStaffIdWorkDateRemarksPatch(body, staffId, workDate, options)(this.fetch, this.basePath);
    }

}